modify double checked locking pattern
